refactor(PlaybookMagicCtrl.js): #16 Shifted blueprint folder cleanup to finally block

+++ src/ctrls/PlaybookMagicCtrl.js
@@ -44,6 +44,7 @@ class PlaybookMagicCtrl extends Controller {
      */
     async convertGit2Playbook(args)
     {
+        let blueprintRepoData;
         try
         {
             /**
@@ -86,55 +87,63 @@ class PlaybookMagicCtrl extends Controller {
             /**
              * @step 2. Create or clone a git repo to store/update the blueprints 
              */
-            let blueprintRepoData = await NodeGitService.createOrCloneBlueprintRepoFromGithubUrl((blueprintGithubUrl ? blueprintGithubUrl : githubUrl), !!blueprintGithubUrl);
+            blueprintRepoData = await NodeGitService.createOrCloneBlueprintRepoFromGithubUrl((blueprintGithubUrl ? blueprintGithubUrl : githubUrl), !!blueprintGithubUrl);
             
-            /**
-             * @step 3. Create the blueprints folder with the github URL
-             */
-            blueprintRepoData = await NodeGitService.createBlueprintsFolderFromGithubUrl(githubUrl, blueprintRepoData);
-
-            /**
-             * @step 4. Ask if we want to build the playbook.json
-             */
-            let buildPlaybookJsonFileAnswer = await inquirer.prompt({
-                type : "confirm",
-                name : "buildPlaybookJsonFile",
-                message : "Would you like to generate the playbook.json file?"
-            })
-
-            const buildPlaybookJsonFile = buildPlaybookJsonFileAnswer.buildPlaybookJsonFile
-
-            if (buildPlaybookJsonFile)
+            if (blueprintRepoData)
             {
-                global.playbook = require('../playbook.sdk').playbook;
-                await PlaybookService.buildPlaybookJsonFromGithub(blueprintGithubUrl, blueprintRepoData);
+                /**
+                 * @step 3. Create the blueprints folder with the github URL
+                 */
+                blueprintRepoData = await NodeGitService.createBlueprintsFolderFromGithubUrl(githubUrl, blueprintRepoData);
+
+                /**
+                 * @step 4. Ask if we want to build the playbook.json
+                 */
+                let buildPlaybookJsonFileAnswer = await inquirer.prompt({
+                    type : "confirm",
+                    name : "buildPlaybookJsonFile",
+                    message : "Would you like to generate the playbook.json file?"
+                })
+
+                const buildPlaybookJsonFile = buildPlaybookJsonFileAnswer.buildPlaybookJsonFile
+
+                if (buildPlaybookJsonFile)
+                {
+                    global.playbook = require('../playbook.sdk').playbook;
+                    await PlaybookService.buildPlaybookJsonFromGithub(blueprintGithubUrl, blueprintRepoData);
+                }
+                else
+                {
+                    process.stderr.write("\nIf you would like to generate the playbook.json then run the command 'playbook build'\n")
+                }
+
+                /**
+                 * @step 5. Push blueprint repo
+                 */
+                await NodeGitService.pushRepo(blueprintRepoData.repo, 
+                                            undefined,
+                                            blueprintRepoData.branch);
+
+                process.stderr.write("\n============ Processing complete ============\n\n")
+                process.stderr.write(`You can view your blueprints at the branch "${blueprintRepoData.branch}" found in the repo:\n`["green"]);
+                process.stderr.write((blueprintRepoData.url + "/tree/" + blueprintRepoData.branch + "\n")["yellow"]);
+                process.stderr.write("\n=============================================\n\n")
             }
-            else
-            {
-                process.stderr.write("\nIf you would like to generate the playbook.json then run the command 'playbook build'\n")
-            }
-
-            /**
-             * @step 5. Push blueprint repo
-             */
-            await NodeGitService.pushRepo(blueprintRepoData.repo, 
-                                          undefined,
-                                          blueprintRepoData.branch);
-
-
-            /**
-             * @step 6. Cleanup blueprint folder
-             */
-            this.deleteFolder(blueprintRepoData.folderPath);
-
-            process.stderr.write("\n============ Processing complete ============\n\n")
-            process.stderr.write(`You can view your blueprints at the branch "${blueprintRepoData.branch}" found in the repo:\n`["green"]);
-            process.stderr.write((blueprintRepoData.url + "/tree/" + blueprintRepoData.branch + "\n")["yellow"]);
-            process.stderr.write("\n=============================================\n\n")
         }
         catch(err)
         {
             console.error(err);
+        }
+        finally
+        {
+            if (blueprintRepoData)
+            {
+                /**
+                 * @step 6. Cleanup blueprint folder
+                 */
+                this.deleteFolder(blueprintRepoData.folderPath);  
+            }
+            
         }
         
         return;
